requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Django==2.1.11
djangorestframework==3.9.3
gunicorn==19.9.0
idna==2.8
Pillow==6.2.0
psycopg2==2.7.5
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
sqlparse==0.3.0
urllib3==1.24.2
apscheduler==3.6.0
python-dotenv==0.11.0
